+++ cms/__init__.py
@@ -3,9 +3,24 @@ from elasticgit import EG
 
 from pyramid_beaker import set_cache_regions_from_settings
 from pyramid.config import Configurator
+from pyramid.i18n import default_locale_negotiator
 
 import logging
 log = logging.getLogger(__name__)
+
+# NOTE
+# Swahili code `swh` is not ISO639-2 so we need to correct this
+# and use `swa` instead.
+LANGUAGE_FALLBACKS = {
+    'swh': 'swa',
+}
+
+# NOTE
+# United Kingdom code `UK` is not ISO3166 so we need to correct this
+# and use `GB` instead.
+COUNTRY_FALLBACKS = {
+    'UK': 'GB',
+}
 
 
 def main(global_config, **settings):
@@ -43,6 +58,32 @@ def init_repository(config):
         log.info('Initialising repository in: %s' % (repo_path,))
 
 
+def get_locale_with_fallbacks(locale_name):
+    if locale_name is None:
+        return None
+
+    language_code, _, country_code = locale_name.partition('_')
+    lang = LANGUAGE_FALLBACKS.get(language_code, language_code)
+    country = COUNTRY_FALLBACKS.get(country_code, country_code)
+
+    if lang != language_code:
+        log.warning(
+            'Invalid language_code used: %s' % language_code,
+            extra={'stack': True})
+
+    if country != country_code:
+        log.warning(
+            'Invalid country_code used: %s' % country_code,
+            extra={'stack': True})
+
+    return u'%s_%s' % (lang, country)
+
+
+def locale_negotiator_with_fallbacks(request):
+    locale_name = default_locale_negotiator(request)
+    return get_locale_with_fallbacks(locale_name)
+
+
 def includeme(config):
     config.include('pyramid_chameleon')
     config.include('pyramid_beaker')
@@ -50,11 +91,13 @@ def includeme(config):
     config.include("pyramid_celery")
     config.add_static_view('static', 'cms:static', cache_max_age=3600)
     config.add_route('home', '/')
+    config.add_route('search', '/search/')
     config.add_route('categories', '/content/list/')
     config.add_route('category', '/content/list/{category}/')
     config.add_route('content', '/content/detail/{uuid}/')
     config.add_route('locale', '/locale/')
     config.add_route('flatpage', '/{slug}/')
     config.scan()
+    config.set_locale_negotiator(locale_negotiator_with_fallbacks)
 
     init_repository(config)