+++ cms/tests/test_views.py
@@ -2,20 +2,54 @@ import arrow
 from datetime import timedelta
 
 from pyramid import testing
-from pyramid.httpexceptions import HTTPNotFound
 from pyramid_beaker import set_cache_regions_from_settings
 
+from cms import locale_negotiator_with_fallbacks
 from cms.tests.base import UnicoreTestCase
 from cms.views.cms_views import CmsViews
 
-from unicore.content.models import Page
+from unicore.content.models import Page, Category, Localisation
 
 
 class TestViews(UnicoreTestCase):
 
     def setUp(self):
         self.workspace = self.mk_workspace()
-        languages = "[('eng_UK', 'English'), ('swh_KE', 'Swahili (Kenya)')]"
+        self.workspace.setup_custom_mapping(Page, {
+            'properties': {
+                'slug': {
+                    'type': 'string',
+                    'index': 'not_analyzed',
+                },
+                'language': {
+                    'type': 'string',
+                    'index': 'not_analyzed',
+                }
+            }
+        })
+        self.workspace.setup_custom_mapping(Category, {
+            'properties': {
+                'slug': {
+                    'type': 'string',
+                    'index': 'not_analyzed',
+                },
+                'language': {
+                    'type': 'string',
+                    'index': 'not_analyzed',
+                }
+            }
+        })
+
+        self.workspace.setup_custom_mapping(Localisation, {
+            'properties': {
+                'locale': {
+                    'type': 'string',
+                    'index': 'not_analyzed',
+                }
+            }
+        })
+
+        languages = "[('eng_GB', 'English'), ('swa_KE', 'Swahili (Kenya)')]"
         settings = {
             'git.path': self.workspace.repo.working_dir,
             'git.content_repo_url': '',
@@ -25,10 +59,13 @@ class TestViews(UnicoreTestCase):
             'cache.long_term.expire': '1',
             'cache.default_term.expire': '1',
             'available_languages': languages,
-            'pyramid.default_locale_name': 'eng_UK',
+            'pyramid.default_locale_name': 'eng_GB',
+            'thumbor.security_key': 'sample-security-key',
+            'thumbor.server': 'http://some.site.com',
         }
         self.config = testing.setUp(settings=settings)
         set_cache_regions_from_settings(settings)
+        self.config.set_locale_negotiator(locale_negotiator_with_fallbacks)
         self.views = CmsViews(testing.DummyRequest())
 
     def tearDown(self):
@@ -67,8 +104,8 @@ class TestViews(UnicoreTestCase):
 
     def test_get_available_languages(self):
         languages = self.views.get_available_languages
-        self.assertEqual(languages[0][0], 'eng_UK')
-        self.assertEqual(languages[1][0], 'swh_KE')
+        self.assertEqual(languages[0][0], 'eng_GB')
+        self.assertEqual(languages[1][0], 'swa_KE')
         self.assertEqual(languages[1][1], 'Swahili (Kenya)')
 
     def test_get_featured_category_pages(self):
@@ -86,31 +123,31 @@ class TestViews(UnicoreTestCase):
 
         self.assertEqual(
             set([page1.language, page2.language]),
-            set(['eng_UK', 'eng_UK']))
+            set(['eng_GB', 'eng_GB']))
 
         self.assertEqual(
             [], list(self.views.get_featured_category_pages(category2.uuid)))
 
     def test_get_featured_category_pages_swahili(self):
-        self.views = CmsViews(testing.DummyRequest({'_LOCALE_': 'swh_KE'}))
+        self.views = CmsViews(testing.DummyRequest({'_LOCALE_': 'swa_KE'}))
 
         [category_eng] = self.create_categories(
-            self.workspace, language='eng_UK', count=1)
+            self.workspace, language='eng_GB', count=1)
         self.create_pages(
-            self.workspace, count=10, language='eng_UK',
+            self.workspace, count=10, language='eng_GB',
             primary_category=category_eng.uuid)
         self.create_pages(
-            self.workspace, count=2, language='eng_UK',
+            self.workspace, count=2, language='eng_GB',
             featured_in_category=True,
             primary_category=category_eng.uuid)
 
         [category_swh] = self.create_categories(
-            self.workspace, language='swh_KE', count=1)
+            self.workspace, language='swa_KE', count=1)
         self.create_pages(
-            self.workspace, count=10, language='swh_KE',
+            self.workspace, count=10, language='swa_KE',
             primary_category=category_swh.uuid)
         pages_swh_featured = self.create_pages(
-            self.workspace, count=2, language='swh_KE',
+            self.workspace, count=2, language='swa_KE',
             featured_in_category=True,
             primary_category=category_swh.uuid)
 
@@ -128,30 +165,16 @@ class TestViews(UnicoreTestCase):
                      category_swh.uuid)]))
 
     def test_get_page_by_slug(self):
-        self.workspace.setup_custom_mapping(Page, {
-            'properties': {
-                'slug': {
-                    'type': 'string',
-                    'index': 'not_analyzed',
-                },
-                'language': {
-                    'type': 'string',
-                }
-            }
-        })
+        self.create_pages(self.workspace, count=5, language='eng_GB')
+        self.create_pages(self.workspace, count=5, language='swa_KE')
 
-        self.create_pages(self.workspace, count=5, language='eng_UK')
-        self.create_pages(self.workspace, count=5, language='swh_KE')
-
-        p = self.views.get_page(None, 'test-page-1', 'eng_UK')
+        p = self.views.get_page(None, 'test-page-1', 'eng_GB')
         self.assertEqual(p.title, 'Test Page 1')
-        self.assertEqual(p.language, 'eng_UK')
+        self.assertEqual(p.language, 'eng_GB')
 
-        p = self.views.get_page(None, 'test-page-1', 'swh_KE')
-        self.assertEqual(p.language, 'swh_KE')
-
-        with self.assertRaises(HTTPNotFound):
-            p = self.views.get_page(None, 'invalid-slug')
+        p = self.views.get_page(None, 'test-page-1', 'swa_KE')
+        self.assertEqual(p.language, 'swa_KE')
+        self.assertEqual(self.views.get_page(None, 'invalid-slug'), None)
 
     def test_content_linked_pages(self):
         [category] = self.create_categories(self.workspace, count=1)
@@ -204,18 +227,6 @@ class TestViews(UnicoreTestCase):
             response['description'], '<p><em>emphasised</em></p>')
 
     def test_flatpage_markdown_rendering(self):
-        self.workspace.setup_custom_mapping(Page, {
-            'properties': {
-                'slug': {
-                    'type': 'string',
-                    'index': 'not_analyzed',
-                },
-                'language': {
-                    'type': 'string',
-                }
-            }
-        })
-
         [category] = self.create_categories(self.workspace, count=1)
         [page] = self.create_pages(
             self.workspace, count=1, content='**strong**',
@@ -232,9 +243,9 @@ class TestViews(UnicoreTestCase):
 
     def test_get_categories(self):
         category1, category2 = self.create_categories(
-            self.workspace, language='eng_UK')
+            self.workspace, language='eng_GB')
         category3, category4 = self.create_categories(
-            self.workspace, language='swh_KE')
+            self.workspace, language='swa_KE')
 
         cat1, cat2 = self.views.get_categories()
         self.assertEqual(
@@ -244,20 +255,42 @@ class TestViews(UnicoreTestCase):
         )
 
         # Change language
-        self.views = CmsViews(testing.DummyRequest({'_LOCALE_': 'swh_KE'}))
+        self.views = CmsViews(testing.DummyRequest({'_LOCALE_': 'swa_KE'}))
         cat1, cat2 = self.views.get_categories()
         self.assertEqual(
             set([cat1.language, cat2.language]),
-            set(['swh_KE', 'swh_KE']))
+            set(['swa_KE', 'swa_KE']))
+
+    def test_categories_ordering(self):
+        category1 = Category(
+            {'title': 'title 1', 'language': 'eng_GB', 'position': 3})
+        category2 = Category(
+            {'title': 'title 2', 'language': 'eng_GB', 'position': 0})
+        category3 = Category(
+            {'title': 'title 3', 'language': 'eng_GB', 'position': 1})
+        category4 = Category(
+            {'title': 'title 4', 'language': 'eng_GB', 'position': 2})
+        self.workspace.save(category1, 'Update position')
+        self.workspace.save(category2, 'Update position')
+        self.workspace.save(category3, 'Update position')
+        self.workspace.save(category4, 'Update position')
+        self.workspace.refresh_index()
+
+        cat1, cat2, cat3, cat4 = self.views.get_categories()
+
+        self.assertEqual(cat1.uuid, category2.uuid)
+        self.assertEqual(cat2.uuid, category3.uuid)
+        self.assertEqual(cat3.uuid, category4.uuid)
+        self.assertEqual(cat4.uuid, category1.uuid)
 
     def test_get_category(self):
         [category] = self.create_categories(
-            self.workspace, count=1, language='swh_KE')
+            self.workspace, count=1, language='swa_KE')
         [page] = self.create_pages(
-            self.workspace, count=1, language='swh_KE',
+            self.workspace, count=1, language='swa_KE',
             primary_category=category.uuid)
 
-        request = testing.DummyRequest({'_LOCALE_': 'swh_KE'})
+        request = testing.DummyRequest({'_LOCALE_': 'swa_KE'})
         request.matchdict['category'] = category.uuid
         views = CmsViews(request)
         response = views.category()
@@ -265,16 +298,150 @@ class TestViews(UnicoreTestCase):
         self.assertEqual(
             [p.uuid for p in response['pages']], [page.uuid])
 
+    def test_pages_ordering(self):
+        [category] = self.create_categories(self.workspace, count=1)
+        page1 = Page({
+            'title': 'title 1', 'language': 'eng_GB', 'position': 3,
+            'primary_category': category.uuid})
+        page2 = Page({
+            'title': 'title 2', 'language': 'eng_GB', 'position': 0,
+            'primary_category': category.uuid})
+        page3 = Page({
+            'title': 'title 3', 'language': 'eng_GB', 'position': 1,
+            'primary_category': category.uuid})
+        page4 = Page({
+            'title': 'title 4', 'language': 'eng_GB', 'position': 2,
+            'primary_category': category.uuid})
+        self.workspace.save(page1, 'Update position')
+        self.workspace.save(page2, 'Update position')
+        self.workspace.save(page3, 'Update position')
+        self.workspace.save(page4, 'Update position')
+        self.workspace.refresh_index()
+
+        request = testing.DummyRequest({})
+        request.matchdict['category'] = category.uuid
+        views = CmsViews(request)
+        cat = views.category()
+        p1, p2, p3, p4 = cat['pages']
+
+        self.assertEqual(p1.uuid, page2.uuid)
+        self.assertEqual(p2.uuid, page3.uuid)
+        self.assertEqual(p3.uuid, page4.uuid)
+        self.assertEqual(p4.uuid, page1.uuid)
+
     def test_get_top_nav(self):
         category1, category2 = self.create_categories(self.workspace)
         category3, category4 = self.create_categories(
-            self.workspace, language='swh_KE', featured_in_navbar=True)
+            self.workspace, language='swa_KE', featured_in_navbar=True)
 
         self.assertEqual([], list(self.views.get_top_nav))
 
         # Change language
-        self.views = CmsViews(testing.DummyRequest({'_LOCALE_': 'swh_KE'}))
+        self.views = CmsViews(testing.DummyRequest({'_LOCALE_': 'swa_KE'}))
         cat1, cat2 = self.views.get_top_nav
         self.assertEqual(
             set([cat1.language, cat2.language]),
-            set(['swh_KE', 'swh_KE']))
+            set(['swa_KE', 'swa_KE']))
+
+    def test_format_date_helper(self):
+        views = CmsViews(testing.DummyRequest({}))
+        self.assertEqual(
+            views.format_date('2014-10-10T09:10:17+00:00'),
+            '10 October 2014')
+
+        self.assertEqual(
+            views.format_date('2014-10-10T09:10:17+00:00', '%d-%b-%y'),
+            '10-Oct-14')
+
+        self.assertEqual(
+            views.format_date('some invalid date'),
+            'some invalid date')
+
+    def test_get_flatpage_using_old_swahili_code(self):
+        [category] = self.create_categories(self.workspace, count=1)
+        [page] = self.create_pages(
+            self.workspace, count=1, content='Sample page in swahili',
+            description='_emphasised_', language='swa_KE')
+
+        request = testing.DummyRequest({'_LOCALE_': 'swh_KE'})
+        request.matchdict['slug'] = page.slug
+        self.views = CmsViews(request)
+        response = self.views.flatpage()
+        self.assertEqual(
+            response['content'], '<p>Sample page in swahili</p>')
+
+    def test_get_flatpage_using_old_english_code(self):
+        [category] = self.create_categories(self.workspace, count=1)
+        [page] = self.create_pages(
+            self.workspace, count=1, content='Sample page in english',
+            description='_emphasised_', language='eng_GB')
+
+        request = testing.DummyRequest({'_LOCALE_': 'eng_UK'})
+        request.matchdict['slug'] = page.slug
+        self.views = CmsViews(request)
+        response = self.views.flatpage()
+        self.assertEqual(
+            response['content'], '<p>Sample page in english</p>')
+
+    def test_image_url(self):
+        self.views = CmsViews(testing.DummyRequest({}))
+
+        self.assertEqual(self.views.get_image_url(
+            'http://some.site.com', 'sample-uuid-000000-0001'),
+            'http://some.site.com/'
+            '1bzRPrcuQPXBECF9mHxFVr11viY=/sample-uuid-000000-0001')
+
+        self.assertEqual(self.views.get_image_url(
+            'http://some.site.com', 'sample-uuid-000000-0001', 300, 200),
+            'http://some.site.com/'
+            '8Ko7ZiKCwOv8zDovqScWL5Lgrc8=/300x200/sample-uuid-000000-0001')
+
+        self.assertEqual(self.views.get_image_url(
+            'http://some.site.com', 'sample-uuid-000000-0001', 300),
+            'http://some.site.com/'
+            'LUyVe1umwB1caELC5m3LWu1HxvI=/300x0/sample-uuid-000000-0001')
+
+        self.assertEqual(self.views.get_image_url(
+            'http://some.site.com', 'sample-uuid-000000-0001', height=150),
+            'http://some.site.com/'
+            '4kS9gT_mYqVhnheDCuQhsahI_dU=/0x150/sample-uuid-000000-0001')
+
+        self.assertEqual(self.views.get_image_url('', ''), '')
+
+    def test_localisation(self):
+        loc = Localisation({
+            'locale': 'eng_GB',
+            'image': 'sample-uuid-000000-0001',
+            'image_host': 'http://some.site.com/'})
+        self.workspace.save(loc, 'Add localisation')
+        self.workspace.refresh_index()
+
+        request = testing.DummyRequest({'_LOCALE_': 'eng_GB'})
+        self.views = CmsViews(request)
+
+        localisation = self.views.get_localisation()
+        self.assertEqual(localisation.locale, 'eng_GB')
+        self.assertEqual(localisation.image, 'sample-uuid-000000-0001')
+        self.assertEqual(localisation.image_host, 'http://some.site.com/')
+
+        # Test fallbacks
+        request = testing.DummyRequest({'_LOCALE_': 'eng_UK'})
+        self.views = CmsViews(request)
+
+        localisation = self.views.get_localisation()
+        self.assertEqual(localisation.locale, 'eng_GB')
+        self.assertEqual(localisation.image, 'sample-uuid-000000-0001')
+        self.assertEqual(localisation.image_host, 'http://some.site.com/')
+
+    def test_localisation_not_found(self):
+        loc = Localisation({
+            'locale': 'eng_GB',
+            'image': 'sample-uuid-000000-0001',
+            'image_host': 'http://some.site.com/'})
+        self.workspace.save(loc, 'Add localisation')
+        self.workspace.refresh_index()
+
+        request = testing.DummyRequest({'_LOCALE_': 'spa_ES'})
+        self.views = CmsViews(request)
+
+        self.assertIsNone(self.views.get_localisation())