+++ cms/views/cms_views.py
@@ -14,7 +14,8 @@ from elasticgit import F
 
 from cms.views.base import BaseCmsView
 
-from unicore.content.models import Category, Page
+from unicore.content.models import Category, Page, Localisation
+from utils import EGPaginator
 
 CACHE_TIME = 'default_term'
 
@@ -31,61 +32,80 @@ class CmsViews(BaseCmsView):
         renderer = get_renderer("cms:templates/base.pt")
         return renderer.implementation().macros['layout']
 
-    def get_categories(self):
-        return self._get_categories(self.locale)
+    @reify
+    def paginator_template(self):
+        renderer = get_renderer("cms:templates/paginator.pt")
+        return renderer.implementation().macros['paginator']
+
+    def get_localisation(self):
+        try:
+            [localisation] = self.workspace.S(
+                Localisation).filter(locale=self.locale)
+            return localisation
+        except ValueError:
+            return None
+
+    def get_categories(self, order_by=('position',)):
+        return self._get_categories(self.locale, order_by)
 
     @cache_region(CACHE_TIME)
-    def _get_categories(self, locale):
-        return self.workspace.S(Category).filter(language=locale.lower())
+    def _get_categories(self, locale, order_by):
+        return self.workspace.S(Category).filter(
+            language=locale).order_by(*order_by)
 
     @cache_region(CACHE_TIME)
     def get_category(self, uuid):
-        [category] = self.workspace.S(Category).filter(uuid=uuid.lower())
+        [category] = self.workspace.S(Category).filter(uuid=uuid)
         return category
 
-    def get_pages(self, limit=5, order_by=('modified_at',)):
+    def get_pages(self, limit=5, order_by=('position', '-modified_at')):
         """
         Return pages the GitModel knows about.
-
         :param int limit:
             The number of pages to return, defaults to 5.
         :param tuple order_by:
-            The attributes to order on, defaults to modified_at
+            The attributes to order on,
+            defaults to ('position', '-modified_at')
         """
         return self.workspace.S(Page).filter(
-            language=self.locale.lower()).order_by(*order_by)[:limit]
+            language=self.locale).order_by(*order_by)[:limit]
 
     @cache_region(CACHE_TIME)
     def _get_featured_pages(self, locale, limit, order_by):
         return self.workspace.S(Page).filter(
-            language=locale.lower(), featured=True).order_by(*order_by)[:limit]
+            language=locale, featured=True).order_by(*order_by)[:limit]
 
-    def get_featured_pages(self, limit=5, order_by=('-modified_at',)):
+    def get_featured_pages(
+            self, limit=5, order_by=('position', '-modified_at')):
         """
         Return featured pages the GitModel knows about.
-
         :param str locale:
             The locale string, like `eng_UK`.
         :param int limit:
             The number of pages to return, defaults to 5.
         :param tuple order_by:
-            The attributes to order on, defaults to ('modified_at',).
+            The attributes to order on,
+            defaults to ('position', '-modified_at').
         """
         return self._get_featured_pages(self.locale, limit, order_by)
 
     @cache_region(CACHE_TIME)
-    def get_pages_for_category(self, category_id, locale):
+    def get_pages_for_category(
+            self, category_id, locale, order_by=('position',)):
         return self.workspace.S(Page).filter(
-            primary_category=category_id, language=locale.lower())
+            primary_category=category_id,
+            language=locale).order_by(*order_by)
 
-    def get_featured_category_pages(self, category_id):
-        return self._get_featured_category_pages(category_id, self.locale)
+    def get_featured_category_pages(
+            self, category_id, order_by=('position',)):
+        return self._get_featured_category_pages(
+            category_id, self.locale, order_by)
 
     @cache_region(CACHE_TIME)
-    def _get_featured_category_pages(self, category_id, locale):
+    def _get_featured_category_pages(self, category_id, locale, order_by):
         return self.workspace.S(Page).filter(
-            primary_category=category_id, language=locale.lower(),
-            featured_in_category=True)
+            primary_category=category_id, language=locale,
+            featured_in_category=True).order_by(*order_by)
 
     @cache_region(CACHE_TIME)
     def get_page(self, uuid=None, slug=None, locale=None):
@@ -93,20 +113,21 @@ class CmsViews(BaseCmsView):
             query = self.workspace.S(Page).filter(
                 F(uuid=uuid) | F(slug=slug))
             if locale is not None:
-                query = query.filter(language=locale.lower())
+                query = query.filter(language=locale)
             [page] = query[:1]
             return page
         except ValueError:
-            raise HTTPNotFound()
+            return None
 
     @reify
-    def get_top_nav(self):
-        return self._get_top_nav(self.locale)
+    def get_top_nav(self, order_by=('position',)):
+        return self._get_top_nav(self.locale, order_by)
 
     @cache_region(CACHE_TIME)
-    def _get_top_nav(self, locale):
+    def _get_top_nav(self, locale, order_by):
         return self.workspace.S(Category).filter(
-            language=locale.lower(), featured_in_navbar=True)
+            language=locale,
+            featured_in_navbar=True).order_by(*order_by)
 
     @view_config(route_name='home', renderer='cms:templates/home.pt')
     @view_config(route_name='categories',
@@ -128,6 +149,10 @@ class CmsViews(BaseCmsView):
     @view_config(route_name='content', renderer='cms:templates/content.pt')
     def content(self):
         page = self.get_page(self.request.matchdict['uuid'])
+
+        if not page:
+            raise HTTPNotFound()
+
         if page.linked_pages:
             linked_pages = self.workspace.S(Page).filter(
                 uuid__in=page.linked_pages)
@@ -150,6 +175,9 @@ class CmsViews(BaseCmsView):
         page = self.get_page(
             None, self.request.matchdict['slug'], self.locale)
 
+        if not page:
+            raise HTTPNotFound()
+
         if page.language != self.locale:
             raise HTTPNotFound()
 
@@ -168,3 +196,41 @@ class CmsViews(BaseCmsView):
                                 value=language,
                                 max_age=31536000)  # max_age = year
         return HTTPFound(location='/', headers=response.headers)
+
+    @view_config(route_name='search', renderer='cms:templates/search.pt')
+    def search(self):
+
+        query = self.request.GET.get('q')
+        p = int(self.request.GET.get('p', 0))
+
+        empty_defaults = {
+            'paginator': [],
+            'query': query,
+            'p': p,
+        }
+
+        # handle query exception
+        if not query:
+            return empty_defaults
+
+        all_results = self.workspace.S(Page).query(content__query_string=query)
+
+        # no results found
+        if all_results.count() == 0:
+            return empty_defaults
+
+        paginator = EGPaginator(all_results, p)
+
+        # requested page number is out of range
+        total_pages = paginator.total_pages()
+        # sets the floor to 0
+        p = p if p >= 0 else 0
+        # sets the roof to `total_pages -1`
+        p = p if p < total_pages else total_pages - 1
+        paginator = EGPaginator(all_results, p)
+
+        return {
+            'paginator': paginator,
+            'query': query,
+            'p': p,
+        }