+++ sideloader/postinstall.sh
@@ -1,4 +1,0 @@
-pip="${VENV}/bin/pip"
-curl https://raw.githubusercontent.com/praekelt/unicore-cms/develop/utils/install_libgit2.sh | sh
-
-$pip install -r "${INSTALLDIR}/${REPO}/requirements.txt"